Contain the integration in the edit-widgets package for now
