+++ packages/edit-widgets/src/store/controls.js
@@ -163,6 +163,7 @@ const controls = {
 				.getEntityRecords( KIND, WIDGET_AREA_ENTITY_TYPE, query );
 		}
 	),
+
 };
 
 const getState = ( registry ) =>
